+++ src/Neo4jBoostServiceProvider.php
@@ -16,6 +16,7 @@ use Neo4j\LaravelBoost\Console\InstallMcpCommand;
 use Neo4j\LaravelBoost\Console\SetupCommand;
 use Neo4j\LaravelBoost\Console\StartNeo4jCommand;
 use Neo4j\LaravelBoost\Console\TestStdioCommand;
+use Neo4j\LaravelBoost\ContainerGraph\AuthConfigExtractor;
 use Neo4j\LaravelBoost\ContainerGraph\BindingLifetimeResolver;
 use Neo4j\LaravelBoost\ContainerGraph\ContextualBindingExtractor;
 use Neo4j\LaravelBoost\ContainerGraph\ContextualGiveResolver;
@@ -91,6 +92,7 @@ class Neo4jBoostServiceProvider extends ServiceProvider
         $this->app->singleton(JobHandlerExtractor::class);
         $this->app->singleton(QueueConnectionExtractor::class);
         $this->app->singleton(ScheduledTaskExtractor::class);
+        $this->app->singleton(AuthConfigExtractor::class);
         $this->app->singleton(LaravelFirstPartyFacadeCatalog::class);
         $this->app->singleton(FacadeCatalogExporter::class);
         $this->app->singleton(FacadeAccessorParser::class);